🐛 fix(none): build app.tsx import issue try-2
